// SDKs/twilio.com/twilio_taskrouter_v1/1.37.3/typescript/src/sdk/models/shared/taskrouterv1workspacetaskqueuetaskqueuestatistics.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";



export class TaskrouterV1WorkspaceTaskQueueTaskQueueStatistics extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @SpeakeasyMetadata({ data: "json, name=cumulative" })
  cumulative?: any;

  @SpeakeasyMetadata({ data: "json, name=realtime" })
  realtime?: any;

  @SpeakeasyMetadata({ data: "json, name=task_queue_sid" })
  taskQueueSid?: string;

  @SpeakeasyMetadata({ data: "json, name=url" })
  url?: string;

  @SpeakeasyMetadata({ data: "json, name=workspace_sid" })
  workspaceSid?: string;
}